# platform/openapi.yaml
openapi: 3.0.3
info:
  title: Platform Service API
  description: |
    User authentication, organization management, invitation system,
    and proxy endpoints to pipeline/plugin microservices.
    Supports JWT and OAuth (Google) authentication.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://platform:3000
    description: Internal Docker network
  - url: http://localhost:3000
    description: Local development

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /organization/{id}/members:
    get:
      summary: List organization members
      operationId: getOrganizationMembers
      tags: [Organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/orgId'
      responses:
        '200':
          description: Members listed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      organizationId: { type: string }
                      organizationName: { type: string }
                      ownerId: { type: string }
                      members:
                        type: array
                        items: { $ref: '#/components/schemas/Member' }
                      total: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    post:
      summary: Add member to organization
      operationId: addMemberToOrganization
      tags: [Organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/orgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: { type: string }
                email: { type: string, format: email }
              description: Provide userId or email (at least one required)
      responses:
        '200':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /organization/{id}/members/{userId}:
    delete:
      summary: Remove member from organization
      operationId: removeMember
      tags: [Organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      summary: Update member role
      operationId: updateMemberRole
      tags: [Organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [user, admin] }
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /organization/{id}/transfer-owner:
    patch:
      summary: Transfer organization ownership
      operationId: transferOrganizationOwnership
      tags: [Organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/orgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newOwnerId]
              properties:
                newOwnerId: { type: string }
      responses:
        '200':
          description: Ownership transferred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /organization/members:
    post:
      summary: Add member to own organization (legacy, org admin)
      operationId: addMember
      tags: [Organization]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /organization/transfer-owner:
    patch:
      summary: Transfer own organization ownership (legacy, org admin)
      operationId: transferOwnership
      tags: [Organization]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newOwnerId]
              properties:
                newOwnerId: { type: string }
      responses:
        '200':
          description: Ownership transferred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------------------------------------------------------------------
  # Invitations
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    refreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT refresh token from login response
  parameters:
    orgId:
      name: id
      in: path
      required: true
      description: Organization ID
      schema: { type: string }
    userId:
      name: id
      in: path
      required: true
      description: User ID
      schema: { type: string }
...
  schemas:
    SuccessResponse:
      type: object
      required: [success, statusCode]
      properties:
        success: { type: boolean, enum: [true] }
        statusCode: { type: integer }
    MessageResponse:
      type: object
      required: [success, statusCode, message]
      properties:
        success: { type: boolean, enum: [true] }
        statusCode: { type: integer }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [success, statusCode, message]
      properties:
        success: { type: boolean, enum: [false] }
        statusCode: { type: integer }
        message: { type: string }
        code: { type: string }
...
    Member:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        role: { type: string, enum: [user, admin] }
        isEmailVerified: { type: boolean }
        isOwner: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - bearerAuth: []
...